+++ Assets/Scripts/ScriptableObjects/CherryPie.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 0}
+  m_Name: CherryPie
+  m_EditorClassIdentifier: Assembly-CSharp::CherryBombItem
+  itemName: Cherry Bomb
+  itemDescription: It explodes.
+  itemValue: 0
+  itemType: 0
+  attackDuration: 0.2
+  attackEndlag: 1
+  attackOffset: {x: 0, y: 0}
+  bomb: {fileID: 1857847825010846, guid: 30614c4eb11041d4385485202196e7e7, type: 2}
+  force: 2.5
