+++ Assets/Scripts/ScriptableObjects/CherryPie.asset.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8ea0b4426edc09f418169bd1cec394cf
+timeCreated: 1507224956
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
